Move middlewares, schemas, utils to common module
